+++ satella/coding/structures/__init__.py
@@ -4,7 +4,7 @@ from .dictionaries import DictObject, apply_dict_object, DictionaryView, TwoWayD
 from .hashable_objects import HashableWrapper
 from .heaps import Heap, SetHeap, TimeBasedHeap, TimeBasedSetHeap
 from .immutable import Immutable, frozendict
-from .mixins import OmniHashableMixin, ReprableMixin
+from .mixins import OmniHashableMixin, ReprableMixin, StrEqHashableMixin
 from .proxy import Proxy
 from .ranking import Ranking
 from .singleton import Singleton, SingletonWithRegardsTo
@@ -16,6 +16,7 @@ __all__ = [
     'KeyAwareDefaultDict',
     'Proxy',
     'ReprableMixin',
+    'StrEqHashableMixin',
     'DirtyDict',
     'SortedList',
     'SelfCleaningDefaultDict',