+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 # v2.8.15
 
+* added for_argument, attach_arguments and auto_adapt_to_methods to __all__ in `satella.coding`
+
+