+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 # v2.8.15
 
 * added for_argument, attach_arguments and auto_adapt_to_methods to __all__ in `satella.coding`
-
-
+* added time_us to __all__ in `satella.time`