+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # v2.8.16
 
-
+* added __slots__ to `StrEqHashableMixin`