+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 * added __slots__ to `StrEqHashableMixin`
 * added `safe_sleep` to `TerminableThread`
+* added `ComparableIntEnum`
+* documentation refactor