+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # v2.8.17
 
+* added make_noncolliding_name
+