+++ CHANGELOG.md
@@ -1,4 +1,2 @@
-# v2.8.17
-
-* added make_noncolliding_name
+# v2.8.18